+++ app/src/main/java/com/mediacontrol/remote/data/WatchRelayListenerService.kt
@@ -1,7 +1,5 @@
 package com.mediacontrol.remote.data
 
-import android.content.Intent
-import androidx.core.content.ContextCompat
 import com.google.android.gms.wearable.DataEventBuffer
 import com.google.android.gms.wearable.WearableListenerService
 import com.mediacontrol.remote.MediaRemoteApp
@@ -10,11 +8,5 @@ import com.mediacontrol.remote.MediaRemoteApp
 class WatchRelayListenerService : WearableListenerService() {
     override fun onDataChanged(dataEvents: DataEventBuffer) {
         (application as MediaRemoteApp).phoneRelay.onDataChanged(dataEvents)
-        if (!UmoBridgeSessionService.isRunning) {
-            ContextCompat.startForegroundService(
-                this,
-                Intent(this, UmoBridgeSessionService::class.java),
-            )
-        }
     }
 }